requirements.txt
# Core LLM + orchestration
crewai>=0.10.0
langchain>=0.1.20
langchain-core>=0.1.46
langchain-openai>=0.1.3
openai>=1.14.0
litellm>=1.35.2

# Web-scraping & search
beautifulsoup4>=4.12.3
requests>=2.31.0
selenium>=4.21.0
webdriver-manager>=4.0.1
tldextract>=5.1.2
duckduckgo-search>=5.2.0

# Email tools
sendgrid>=6.11.0
dnspython>=2.4.2

# Excel / data
pandas>=2.2.1
openpyxl>=3.1.2

# App layer
streamlit>=1.34.0
python-dotenv>=1.0.1
python-decouple>=3.8

# Config / parsing
PyYAML>=6.0.1

# File watching (needed for Streamlit local hot reload)
watchdog>=4.0.0
